+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/SpringBooter.java
@@ -14,13 +14,19 @@
  */
 package com.ericsson.bss.cassandra.ecchronos.application;
 
+import com.ericsson.bss.cassandra.ecchronos.rest.OnDemandRepairManagementRESTImpl;
+import com.ericsson.bss.cassandra.ecchronos.rest.RepairManagementRESTImpl;
+import com.ericsson.bss.cassandra.ecchronos.rest.ScheduleRepairManagementRESTImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
+@Import(value = { RepairManagementRESTImpl.class, ScheduleRepairManagementRESTImpl.class,
+        OnDemandRepairManagementRESTImpl.class })
 public class SpringBooter extends SpringBootServletInitializer
 {
     private static final Logger LOG = LoggerFactory.getLogger(SpringBooter.class);